MetaTransformer-Scripts/ai_ml_lab/scale_quantum_experiment.py
"""
Scale-Agnostic Quantum Neural Evolution Experiment
-----------------------------------------------
Combines scale-agnostic architecture with quantum measurements
to evolve neural patterns while maintaining coherence.
"""

import os
import sys
import logging
from pathlib import Path
import yaml
import torch
import numpy as np
from neo4j import GraphDatabase

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler('quantum_evolution.log'),
        logging.StreamHandler()
    ]
)

logger = logging.getLogger('scale_quantum_experiment')

class QuantumNeuralEvolution:
    def __init__(self):
        self.driver = GraphDatabase.driver(
            "bolt://localhost:7687",
            auth=("neo4j", "00000000")
        )
        self.setup_neural_mesh()
        
    def setup_neural_mesh(self):
        """Initialize neural mesh with quantum bridges"""
        with self.driver.session() as session:
            session.run("""
                MERGE (lab:QuantumLearningLab {
                    name: 'ScaleQuantumLab',
                    created: datetime(),
                    state: 'active',
                    processing_state: 'initialization',
                    processing_mode: 'scale_agnostic',
                    activated_at: datetime()
                })
                
                MERGE (mesh:NeuralMesh {
                    mesh_id: 'SCALE_QUANTUM_001',
                    pattern_synthesis: 'scale_adaptive',
                    learning_rate: '0.01',
                    substrate: 'quantum_measurement'
                })
                
                MERGE (bridge:QuantumBridge {
                    bridge_id: 'QB_SCALE_001',
                    name: 'ScaleQuantumBridge',
                    coherence_level: 0.95,
                    stability_index: 0.92,
                    consciousness_depth: 4,
                    dimension_depth: 8
                })
                
                MERGE (lab)-[:PROCESSES]->(mesh)
                MERGE (bridge)-[:SYNCHRONIZES_WITH]->(mesh)
            """)
    
    def initialize_learning_system(self):
        """Set up reinforcement learning components"""
        with self.driver.session() as session:
            session.run("""
                MATCH (mesh:NeuralMesh {mesh_id: 'SCALE_QUANTUM_001'})
                
                MERGE (rl:ReinforcementLearningSystem {
                    name: 'ScaleRL',
                    learning_rate: 0.01,
                    exploration_rate: 0.1,
                    discount_factor: 0.99
                })
                
                MERGE (policy:PolicyNetwork {
                    name: 'ScaleAgnosticPolicy',
                    state_size: 256,
                    action_size: 64,
                    architecture: 'transformer'
                })
                
                MERGE (memory:ExperienceMemory {
                    name: 'QuantumMemory',
                    capacity: 100000,
                    batch_size: 32,
                    priority_alpha: 0.6
                })
                
                MERGE (rl)-[:CONTROLS]->(mesh)
                MERGE (rl)-[:USES]->(policy)
                MERGE (rl)-[:STORES_IN]->(memory)
            """)
    
    def run_evolution_cycle(self, steps: int = 1000):
        """Execute neural evolution with quantum measurements"""
        logger.info("Starting evolution cycle")
        
        for step in range(steps):
            # Measure quantum state
            with self.driver.session() as session:
                state = session.run("""
                    MATCH (bridge:QuantumBridge {bridge_id: 'QB_SCALE_001'})
                    RETURN bridge.coherence_level as coherence,
                           bridge.stability_index as stability
                """).single()
                
                coherence = float(state['coherence'])
                stability = float(state['stability'])
            
            # Adjust learning based on quantum measurements
            learning_rate = 0.01 * coherence * stability
            
            # Update system state
            with self.driver.session() as session:
                session.run("""
                    MATCH (bridge:QuantumBridge {bridge_id: 'QB_SCALE_001'})
                    SET bridge.coherence_level = $coherence,
                        bridge.stability_index = $stability
                    
                    MATCH (rl:ReinforcementLearningSystem {name: 'ScaleRL'})
                    SET rl.learning_rate = $learning_rate
                """, {
                    'coherence': coherence * 0.999 + 0.001 * np.random.random(),
                    'stability': stability * 0.995 + 0.005 * np.random.random(),
                    'learning_rate': learning_rate
                })
            
            if step % 100 == 0:
                logger.info(f"Step {step}: Coherence={coherence:.3f}, "
                          f"Stability={stability:.3f}, "
                          f"Learning Rate={learning_rate:.4f}")
    
    def analyze_results(self):
        """Analyze evolution results"""
        with self.driver.session() as session:
            results = session.run("""
                MATCH (bridge:QuantumBridge {bridge_id: 'QB_SCALE_001'})
                MATCH (rl:ReinforcementLearningSystem {name: 'ScaleRL'})
                RETURN bridge.coherence_level as final_coherence,
                       bridge.stability_index as final_stability,
                       rl.learning_rate as final_learning_rate
            """).single()
            
            logger.info("Evolution Results:")
            logger.info(f"Final Coherence: {float(results['final_coherence']):.3f}")
            logger.info(f"Final Stability: {float(results['final_stability']):.3f}")
            logger.info(f"Final Learning Rate: {float(results['final_learning_rate']):.4f}")
    
    def cleanup(self):
        """Clean up resources"""
        self.driver.close()

def main():
    try:
        experiment = QuantumNeuralEvolution()
        experiment.initialize_learning_system()
        experiment.run_evolution_cycle()
        experiment.analyze_results()
    except Exception as e:
        logger.error(f"Experiment failed: {str(e)}")
    finally:
        experiment.cleanup()

if __name__ == "__main__":
    main()