MetaTransformer-Scripts/ai_ml_lab/requirements.txt
torch>=2.1.0
numpy>=1.24.0
neo4j>=5.15.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
logging>=0.5.1.2